Code cleaning: convert the DIALOG_PASS macro into a regular function.
This is the first commit of a serie, aimed to improve readability and to reduce
code duplication of the dialog system.

The DIALOG_PASS macro had a lot of assumptions on the variables (and their
types) that must exist when called. The replacement by a function forces theses
variables to be explicitely passed as argument.

This commit converts `DIALOG_PASS` into a new function `dialog_pass()`, located
in the new file `newGui/Dialog.cpp`. Also, some doc has been added.

The content of the function itself is unmodified.

..