Code cleaning: create a `Dialog` class from duplicated code
There is 3 dialogs with duplicated code, located in `Startscene.cpp`
(`boucle0()` and `boucle2`) and in `Overworld.cpp` (`boucleDialog()`). Each of
these methods handles a game loop, the display phase, and a dialog details (text
animation, draw, ...).
This commit moves everything related to the dialog into a `Dialog` class.

The content of the code itself remains almost unchanged.