Fix returning of resource path
We are in a while loop and plan to go through all elements of
XDG_DATA_DIRS. So far we actually returned out of it if only the first
path is not containig the OpMon dir, not even checking the other
elements.

Only return the default /usr/local/share/OpMon path if none of the
elements in XDG_DATA_DIRS contain OpMon.

Fixes https://github.com/OpMonTeam/OpMon/issues/52